Add events to interact with cells.
Now when you click on a cell, it shows an input.
If the user types a number, the last non-zero number is remembered.
If they type enter, we remember the input.
If they go away from the input, we go away from editing.

If the user its escape, we revert back to the value it was.